+++ CHANGELOG.md
@@ -13,6 +13,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - renamed variable `unicorn_binance_websocket_api_manager` to `manager` in socket class
 - renamed variable `total_receiving_speed` to `average_receiving_speed`
 - user agent string for rest and websocket client
+### Fixed
+- `TypeError` exception in `_frequent_checks()`
  
 ## 1.21.0
 ### Added